[更新double =》 int]

+++ src/main/java/com/seblong/okr/services/OKRService.java
@@ -14,7 +14,7 @@ public interface OKRService {
 
 	Objective addObjective(String user, String period, String title);
 
-	Objective updateObjective(String user, String period, String id, String title, long estimate, double confidence)
+	Objective updateObjective(String user, String period, String id, String title, long estimate, int confidence)
 			throws ValidationException;
 
 	Objective deleteObjective(String user, String period, String id);
@@ -23,13 +23,13 @@ public interface OKRService {
 
 	KeyResult addKeyResult(String user, String period, String objective, String title) throws ValidationException;
 
-	KeyResult updateKeyResult(String user, String period, String objective, String id, String title,long estimate, double confidence, double weight, double progress) throws ValidationException;
+	KeyResult updateKeyResult(String user, String period, String objective, String id, String title,long estimate, int confidence, int weight, int progress) throws ValidationException;
 
 	KeyResult deleteKeyResult(String user, String period, String objective, String id);
 
 	void updateProgress(String user, String period, String objective, String progress) throws ValidationException;
 
-	Objective scoreKeyResult(String user, String period, String objective, String id, double score)
+	Objective scoreKeyResult(String user, String period, String objective, String id, int score)
 			throws ValidationException;
 	
 	void rankKeyResult(String user, String period, String objective, int oldRank, int newRank);
